+++ runtime/src/Makefile.share
@@ -41,6 +41,7 @@ COMMON_NOGEN_SRCS = \
 	chpl-mem-desc.c \
 	chpl-mem-hook.c \
 	chplmemtrack.c \
+	chpl-privatization.c \
 	chpl-string.c \
 	chplsys.c \
 	chpl-tasks.c \
